+++ src/grid.c
@@ -67,7 +67,7 @@ float **matrixreserv(int nrl, int nrh, int ncl, int nch)
 /*   fuer DOUBLE :  aus Numerical Receipes in C               */
 /**************************************************************/
 
-double **matrixreservdouble(long nrl, long nrh, long ncl, long nch)
+double **matrixreservdouble(int nrl, int nrh, int ncl, int nch)
 
 {
     /*nrl = number of rows lower limit, nrh = high limit, c = columns*/
